+++ KeySwitchManager/Sources/Runtime/Applications/CLI/Sources/Commands/New.cs
@@ -22,9 +22,7 @@ namespace KeySwitchManager.Applications.CLI.Commands
             var option = (CommandOption)opt;
             var logView = new ConsoleLogView();
 
-            ICreateFileControllerFactory factory = new CreateFileControllerFactory();
-
-            using var controller = factory.Create( option.OutputPath, new CreatePresenter( new ConsoleLogView() ) );
+            IController controller = CreateController.Create( option.OutputPath, new CreatePresenter( new ConsoleLogView() ) );
             logView.Append( $"generating keyswitch template to {option.OutputPath}" );
             controller.Execute();
