--- KeySwitchManager/Sources/Runtime/Applications/Standalone.Core/KeySwitches/Controllers/Create/CreateFileControllerFactory.cs
@@ -1,43 +1,0 @@
-﻿using System;
-
-using KeySwitchManager.Commons.Data;
-using KeySwitchManager.Controllers.KeySwitches;
-using KeySwitchManager.Infrastructures.Storage.KeySwitches;
-using KeySwitchManager.Infrastructures.Storage.Spreadsheet.ClosedXml.KeySwitches.Export;
-using KeySwitchManager.Infrastructures.Storage.Yaml.KeySwitches.Export;
-using KeySwitchManager.UseCase.KeySwitches.Create;
-using KeySwitchManager.UseCase.KeySwitches.Export;
-
-namespace KeySwitchManager.Applications.Standalone.Core.KeySwitches.Controllers.Create
-{
-    public class CreateFileControllerFactory : ICreateFileControllerFactory
-    {
-        IController ICreateFileControllerFactory.Create( string outputFilePath, ICreatePresenter presenter )
-        {
-            var fileName = outputFilePath.ToLower();
-
-            IExportContentFactory contentFactory;
-            IExportContentWriterFactory contentWriterFactory;
-            IExportStrategy strategy;
-
-            if( fileName.EndsWith( ".xlsx" ) )
-            {
-                contentFactory       = new ClosedXmlExportContentFactory();
-                contentWriterFactory = new ClosedXmlExportContentFileWriterFactory( new SpecificExportPathBuilder( new FilePath( outputFilePath ) ) );
-                strategy             = new SingleExportStrategy( contentWriterFactory, contentFactory );
-            }
-            else if( fileName.EndsWith( ".yaml" ) || fileName.EndsWith( ".yml" ) )
-            {
-                contentFactory       = new YamlExportContentFactory();
-                contentWriterFactory = new YamlExportContentFileWriterFactory( new SpecificExportPathBuilder( new FilePath( outputFilePath ) ) );
-                strategy             = new SingleExportStrategy( contentWriterFactory, contentFactory );
-            }
-            else
-            {
-                throw new ArgumentException( $"{outputFilePath} is unknown file format" );
-            }
-
-            return new CreateFileController( strategy, presenter );
-        }
-    }
-}
